restructured app client and db to be at top level homescreen, removed error of setting stitch.defaultappclient more than once
